chore: create .gitlab-ci.yml pipeline
The pipeline checks constraints, packages and symlinks.
It builds the packages on the workspace and tests them.

The first version of the deploy script creates a cloudformation
stack with cloudfront and an s3 bucket to every pull request.
The content of the apps/frontend build folder is synced to the
bucket.