fix(page-controls): '1 ...' on mobile
If there are two pages, and the current page is the first,
'1 ...' was shown on mobile, while on the second page '1 2' was shown
correctly. This fixes it so that both pages show '1 2' correctly.

Unfortunately, this cannot be tested yet, because JSDOM doesn't let us
check how things look on mobile. I may have something in the pipeline
for that...