perf(frontend): prefetch more sections
And prefetch sequentially. This means that on slower networks (where
prefetching matters most, and the bandwidth is so constrained that
fetching more bundles at once does not increase overall throughput), the
first sections will complete prefetching earlier because they do not
share bandwidth with other bundles being prefetched. The order should
reflect how likely we think users are to navigate to each section.