fix: team header Avatar scaling
I've changed the team page to have enough space available.

Avatars can now scale according to the available space.

I hadn't considered the difficulty scaling the initials,
but with a simple SVG it's reasonably easy,
and the autoscaling has e.g. the advantage of being able to
fill exactly one row with Avatars, as is done in the new team header.

I've also refactored the team header's extra users bubble
into the Avatar component to share styling with the real Avatar bubbles.