fix(modals): prevent navigation when already elsewhere
We first tried actually aborting the save requests,
however that would almost 100% of the time not actually prevent the
change being made, meaning now we had to initiate another GET for the
changed value. This wastes resources and increased the code of each
useXById and their tests by a lot.

We decided to instead let the request complete (and update our state
with the change) and only ensure that the Promise resolution does not
trigger a navigation when the user might already be somewhere else
entirely.

There is technically another edge case now when you save-abort-reenter
the modal, but the incoming response closing the modal in that case
actually seems reasonable given that the modal contains stale data.