refactor(react-components): remove brittle automatic required asterisks
Some places don't want the asterisk. The signin page, for example.
There, it only works (doesn't show) so far because the email field happens to not be required
(it does not matter whether it's set to required because it's type="email"),
and because password fields do not have the automatic asterisk logic yet.