CRN-1191 - Automatically update deliverable statuses on working groups
When changing the complete property of a working group the status of deliverables will also update so that complete working groups have "Incomplete" and "Not Started" deliverables where incomplete working group deliverables are "In Progress" and "Pending" respectively.

Adds an eventbridge handler listening for `WorkingGroupUpdate` and `WorkingGroupCreate` events and updates the deliverables accordingly.