SPIKE: allow adding a label to a PR to completely skip CI
For some PRs that are work in progress, or shouldn't run full CI for some other reason support adding a label to the PR to skip over the entire CI pipeline.

Note that this _fails_ the CI, so it shouldn't be possible to accidentally merge the PR with that label, and eventually it will need to be removed and run once before merging.