CRN-1336 - configure fail on console to ignore recoil warnings (#2799)
* CRN-1336 - configure fail on console to ignore recoil warnings

And remove the retry, which I understand to be an attempt to brute force past those warnings.

* Improvement: remove final pipeline level retry

This suite is pretty reliable so shouldn't need a retry. If it becomes flaky then fix the flake rather than brute-forcing.

* CRN-1336 - update jest-fail-on-console to account for group context

Allows us to silence console warnings based on the `console.group` context in which they appear, and so allows ignoring recoil cache key warnings in tests.