Take expected timestamp before request (#3330)
Refactor the integration tests for setting created and updated timestamps to set the expectation as the date before the request is made. This means that if there are long waits for the request to resolve (due to contentful rate limiting) then the expected timestamp should still be within the expected bounds.