Don't strip non-empty top-level text nodes (#3537)
the migration was removing all text nodes that were at the top level of the rich text structure, irrespective of whether they contained text. This resulted in the loss of text data from the description fields of research outputs.

Instead, where there are top-level text nodes that contain text, wrap them into a paragraph node and preserve the content.