Use Content Delivery API for polling Contentful
The caching rules for the Content Delivery API and Content Preview API are different, so a record returning as being up to date against the CPA does not guarantee that a subsequent GraphQL request will return fresh data.

Using the Content Delivery API, which seems to use the same caching rules as GraphQL, ensures that the record is up to date before emitting to EventBridge.