Handle errors in algolia search requests
If algolia returns an error while searching then it gets stuck in a continuous loop of retrying the request. This is particularly bad when the error is a 429 response for making too many requests.

Instead of getting stuck in a continuous loop of retrying failed searches throw an error and display a failure message.