Don't add draft research outputs to algolia
The code previously assumed that if a research output was unpublished then fetching the output would result in a "Not Found" error, but since the concept of draft outputs was introduced this is no longer the case, and the unpublish case needs to be handled separately from the deleted/nout found case.

If a research output is not published then remove it from the algolia index.