Fix removal of objects from Algolia when they are unpublished
The `fetchById` methods on the controllers throw `NotFoundError` when an entity is unpublished, but the handlers expect a `Boom` instance. This means that they do not remove entities from Algolia when they are removed from the CMS.

Additionally handle the `NotFoundError` case in the user, event, and research output indexing handlers so they are correctly deleted from Algolia when unpublished.