Always sync settings when indexing algolia
Running a partial index was excluding the settings from the scope of the index copy, which meant that the temporary index would not have any searchable attributes, and when moved back to replace the main index would therefore remove any searchable attributes.

Remove the scope parameter from partial indexing runs so that all settings and data is preserved.