Allow more complex queries for single entities than lists (#3594)
Fetching all the data for an entity on the list endpoint often results in exceeding query complexity limits in Contentful where there are multiple related entities. In particular for relted research authors and event speakers.

Where these are not rendered into the list view, or used for searching this can be solved by only loading these relations when querying for a single entity, and not for list endpoints.

Add an extra recoil entity for the list representations of entities so that the full data is always fetched from the server when clicking from a list entry into the full detail page.