Ensure uniqueness of channel ids when subscribing to google calendars
It is not possible to create multiple subscriptions with the same channel id, but the resource id returned when creating a subscription is required to unsubscribe.

This means that if the resource id is removed from the record in the CMS, then a new subscription cannot be created because the channel id is non unique, but the previous subscription cannot be removed.

Ensure that new subscriptions can always be created by adding a timestamp onto the channel id to ensure uniqueness. Also store the generated channel id when subscribing so that it can be used if needed to unsubscribe in future.