Map empty string values in user profile to null
Passing an empty string for a non-required field with pattern or "allowed values" validators throws an error. In particular when onboarding a new user, the "Degree" field is marked as optional on the frontend, but not selecting a value causes an error because `""` is not one of the allowed vlus in Contentful.

Map any empty string values to `null` in the data provider to prevent the validator throwing an error.